+++ grafana-plugin/src/containers/CreateAlertReceiveChannelContainer/CreateAlertReceiveChannelContainer.tsx
@@ -66,7 +66,7 @@ const CreateAlertReceiveChannelContainer = observer((props: CreateAlertReceiveCh
       <div className={cx('select-team')}>
         <Field
           label="Assign to team"
-          description="When an integration is assigned to a team, all the produced alerts will belong to this team"
+          description="OnCall teams allow you to organize integrations so you can filter and set up access. "
         >
           <GrafanaTeamSelect withoutModal onSelect={setSelectedTeam} />
         </Field>
@@ -93,7 +93,9 @@ const CreateAlertReceiveChannelContainer = observer((props: CreateAlertReceiveCh
                 </div>
                 <div className={cx('title')}>
                   <VerticalGroup spacing="none">
-                    <Text strong>{alertReceiveChannelChoice.display_name}</Text>
+                    <Text strong data-testid="integration-display-name">
+                      {alertReceiveChannelChoice.display_name}
+                    </Text>
                     <Text type="secondary" size="small">
                       {alertReceiveChannelChoice.short_description}
                     </Text>