+++ grafana-plugin/src/containers/OutgoingWebhookForm/OutgoingWebhookForm.config.ts
@@ -11,6 +11,8 @@ export const form: { name: string; fields: FormItem[] } = {
     {
       name: 'team',
       label: 'Assign to team',
+      description:
+        'Assigning to the teams allows you to filter Outgoing Webhooks and configure their visibility. Go to OnCall -> Settings -> Team and Access Settings for more details',
       type: FormItemType.GSelect,
       extra: {
         modelName: 'grafanaTeamStore',