+++ grafana-plugin/src/pages/integrations/Integrations.tsx
@@ -149,7 +149,12 @@ class Integrations extends React.Component<IntegrationsProps, IntegrationsState>
           <>
             <div className={cx('root')}>
               <div className={cx('filters')}>
-                <RemoteFilters query={query} page="integrations" onChange={this.handleIntegrationsFiltersChange} />
+                <RemoteFilters
+                  query={query}
+                  page="integrations"
+                  grafanaTeamStore={store.grafanaTeamStore}
+                  onChange={this.handleIntegrationsFiltersChange}
+                />
               </div>
               {data?.length ? (
                 <div className={cx('integrations')}>