+++ grafana-plugin/src/pages/schedules/Schedules.tsx
@@ -152,7 +152,12 @@ class SchedulesPage extends React.Component<SchedulesPageProps, SchedulesPageSta
         <div className={cx('root')}>
           <VerticalGroup>
             <div className={cx('schedules__filters-container')}>
-              <RemoteFilters query={query} page="schedules" onChange={this.handleSchedulesFiltersChange} />
+              <RemoteFilters
+                query={query}
+                page="schedules"
+                grafanaTeamStore={store.grafanaTeamStore}
+                onChange={this.handleSchedulesFiltersChange}
+              />
               <div className={cx('schedules__actions')}>
                 {users && (
                   <UserTimezoneSelect
@@ -311,8 +316,8 @@ class SchedulesPage extends React.Component<SchedulesPageProps, SchedulesPageSta
                   relatedEscalationChains.length ? (
                     relatedEscalationChains.map((escalationChain) => (
                       <div key={escalationChain.pk}>
-                        <PluginLink query={{ page: 'escalations', id: escalationChain.pk }}>
-                          {escalationChain.name}
+                        <PluginLink query={{ page: 'escalations', id: escalationChain.pk }} className="link">
+                          <Text type="link">{escalationChain.name}</Text>
                         </PluginLink>
                       </div>
                     ))
@@ -331,7 +336,6 @@ class SchedulesPage extends React.Component<SchedulesPageProps, SchedulesPageSta
         {item.warnings?.length > 0 && (
           <ScheduleCounter
             type="warning"
-            addPadding
             count={item.warnings.length}
             tooltipTitle="Warnings"
             tooltipContent={